refactor: Address input select (#2210)
